remove old generic mutation functions, fix examples
