+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Entities can be initialized with components, via ID as well as using generics (#76)
 * A listener function can be registered to the world, for notification on entity changes (#77)
 * Support for up to 128 distinct component types per world (was limited to 64 before) (#78)
+* Generic entity manipulation through types `Mutate1`, `Mutate2`, ... (#79)
 
 ### Other
