+++ benchmark/add/arche_test.go
@@ -26,10 +26,11 @@ func addArcheGeneric(b *testing.B, count int) {
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
 		world := ecs.NewWorld()
+		mut := generic.NewMutate2[position, rotation](&world)
 		b.StartTimer()
 
 		for i := 0; i < count; i++ {
-			_, _, _ = generic.NewEntity2[position, rotation](&world)
+			_, _, _ = mut.NewEntity()
 		}
 	}
 }
